+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.1.0 (unreleased)
 ------------------
 
+- #7 Do not allow microorganisms with same title
 - #6 Remove IMicroorganismBehavior and make Microorganism folderish